- bug fix for restart scripts.

+++ capstone/launchOrderer.sh
@@ -32,8 +32,13 @@ fi
 pids=`ps ax | grep -i 'orderer' | grep -v grep | awk '{print $1}'`
 for pid in $(eval echo $pids)
 do
-	echo "########### kill -9 $pid #############"
-	kill -9 $pid
+    if [ $pid == $$ ]
+    then
+        echo "Current PID = " $pid
+    else
+       echo "########### kill -9 $pid #############"
+       kill -9 $pid
+    fi
 done
 
 rm -rf $HOME/orderer
